Added: Timestamp of the application is stored appropriately. Not implemented for the convertDb.py script yet (will do next) Fixed: Now using int64_t, as SQLite is not able to store bigger values (or uint64_t)

+++ src/cxxprof_dyn_network/AppInfo.h
@@ -4,6 +4,7 @@
 
 #include "cxxprof_dyn_network/common.h"
 #include <map>
+#include <stdint.h>
 #include <string>
 
 namespace CxxProf
@@ -12,7 +13,7 @@ namespace CxxProf
     struct CxxProf_Dyn_Network_EXPORT AppInfo
     {
         std::string Name;
-        unsigned int Starttime;
+        int64_t Starttime;
         std::map<unsigned int, std::string> ThreadAliases;
 
         AppInfo() :
